文件审核修改

+++ src/operations/admin.ts
@@ -109,25 +109,35 @@ Admin.permitFile = async function permitFile(fileId: string): Promise<boolean> {
   const con = await basic('cloud');
   const sql = `select * from pending_file where id=${fileId}`;
   const result = await Query(sql, con);
-  const delSql = `delete from pending_file where id=${fileId}`;
-  await Query(delSql, con);
-  const addSql = `insert into file(id,filename,type,size,downloads,hash) values(${
-    result[0].id
-  },'${result[0].filename}','${result[0].type}',${result[0].size},${0},'${
-    result[0].hash
-  }')`;
-  await Query(addSql, con);
-  return true;
+  if (result.length) {
+    const delSql = `delete from pending_file where id=${fileId}`;
+    await Query(delSql, con);
+    const addSql = `insert into file(id,filename,type,size,downloads,hash) values(${
+      result[0].id
+    },'${result[0].filename}','${result[0].type}',${result[0].size},${0},'${
+      result[0].hash
+    }')`;
+    await Query(addSql, con);
+    return true;
+  } else {
+    return false;
+  }
 };
 
 Admin.rejectFile = async function permitFile(fileId: string): Promise<boolean> {
   const con = await basic('cloud');
-  const delSql = `delete from pending_file where id=${fileId}`;
-  await Query(delSql, con);
-  return true;
+  const sql = `select * from pending_file where id=${fileId}`;
+  const result = await Query(sql, con);
+  if (result.length) {
+    const delSql = `delete from pending_file where id=${fileId}`;
+    await Query(delSql, con);
+    return true;
+  } else {
+    return false;
+  }
 };
 
-Admin.getPendingFiles = async function getPermitFiles() {
+Admin.getPendingFiles = async function getPermitFiles(): Promise<{}> {
   const con = await basic('cloud');
   const sql = `select * from pending_file`;
   const result = await Query(sql, con);
